server/requirements.txt
absl-py==1.4.0
beautifulsoup4==4.12.2
CacheControl==0.12.11
deepdiff==8.6.1
Flask[async]==2.3.2
Flask-Babel==2.0.0
Flask-Caching==2.0.1
flask_cors==5.0.0
flask_testing==0.8.1
frozendict==2.3.4
geojson_rewind==1.0.1
fastapi==0.116.0
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
google-adk==1.18.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-api-core>=2.18.0,<3.0.0
google-cloud-logging>=3.11.0,<4.0.0
google-cloud-aiplatform==1.125.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.1
google-cloud-discoveryengine==0.13.12
google-cloud-language
google-cloud-secret-manager==2.22.0
google-cloud-storage==3.0.0
google-genai==1.45.0
gunicorn==23.0.0
isort==5.10.0
jinja2==3.1.6
json5==0.9.14
langdetect==1.0.9
markupsafe==2.1.2
numpy==2.2.4
parameterized==0.8.1
pillow==10.4.0
protobuf>=4.25.3
pydantic==2.11.1
PyGithub==1.58.2
pyOpenSSL==23.2.0
python-dateutil>=2.9.0.post0, <3.0.0
pytest-flakefinder==1.1.0
pytest-rerunfailures==10.2
pytest-recording==0.13.2
pytest-xdist==3.2.1
PyYAML==6.0.2
redis==4.5.4
requests==2.32.4
selenium==4.21.0
typing-extensions==4.12.2
webdriver-manager==4.0.0
Werkzeug==3.0.6
wheel==0.38.1
# TODO: publish the library to pypi so we don't have to install from git
git+https://github.com/datacommonsorg/llm-tools.git